- Put enum localbool inside a #ifdef so that it's only compiled on FreeBSD.

- Removed unused varibles.

- Got rid of extra -lpthread in the Makefile.

- Include openssl/rand.h so that crypto compiles on Mac OSX.

- Switch to c99 from c11 because c11 support is terrible outside of FreeBSD and   presumely Linux.

- Turned on -Werror and -pendantic

- Added newlines to the end of source files.

- Added pragmas to deps so they want emit warnings.

- Imported Splint for improved testing and code quality. This splint has minor
  fixes so that it compiles on Mac OSX.

- Removed a few unneeded header includes.

- Imported LibreSSL into project so that Mac OSX user's are not forced to
  download openssl or libressl on 10.11 El Capitan which is no longer included.
  This also forces user's on other platforms to use a more sane crypto library.

- Switch to pthreads from standard c11 threads because Mac OSX does not support
  c11 threads.

- Added arg_type_ctx so that large switch staments can be eliminated when
  dealing with syscall argument types.

+++ src/io.c
@@ -147,7 +147,7 @@ int copy_file_to(char *src, char *dst)
         return -1;
     }
 
-    munmap(file_buffer, file_size);
+    munmap(file_buffer, (size_t)file_size);
 
     return 0;
 }
